split up fileTypes test
