+++ packages/type-bmp/package.json
@@ -5,6 +5,9 @@
   "main": "dist/index.js",
   "module": "es/index.js",
   "scripts": {
+    "test": "cross-env BABEL_ENV=test mocha --require @babel/register",
+    "test:watch": "npm run test -- --reporter min --watch",
+    "test:coverage": "nyc npm run test",
     "build": "npm run build:node:production && npm run build:module",
     "build:watch": "npm run build:node:debug -- -- --watch --verbose",
     "build:debug": "npm run build:node:debug",
@@ -22,6 +25,9 @@
   },
   "peerDependencies": {
     "@jimp/custom": ">=0.3.5"
+  },
+  "devDependencies": {
+    "@jimp/test-utils": "^0.3.9"
   },
   "publishConfig": {
     "access": "public"
