+++ packages/type-bmp/test/images/windows95.png
