+++ packages/type-jpeg/test/images/fillbytes.jpg
