+++ packages/type-png/babel.config.js
@@ -1,0 +1,1 @@
+../../babel.config.js
